fixed bugs in save/load file system + position is now double, but saves as int.

+++ src/main/java/cz/cvut/copakond/pinkfluffyunicorn/model/entities/Character.java
@@ -10,7 +10,7 @@ public class Character extends GameObject implements ICharacter {
     boolean isEnemy;
     String name;
 
-    public Character(String textureName, int[] position, DirectionEnum direction) {
+    public Character(String textureName, double[] position, DirectionEnum direction) {
         super(textureName, position, RenderPriorityEnums.CHARACTER.getValue());
         this.direction = direction;
         this.name = textureName;
